requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-appconf==1.0.5
django-auth-ldap==4.3.0
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-select2==8.1.2
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
lxml==4.9.3
MarkupPy==1.14
MarkupSafe==2.1.3
ms-identity-web==0.16.6
msal==1.23.0
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
Pillow==10.0.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pypdf==3.12.2
pypng==0.20220715.0
python-bidi==0.4.2
python-ldap @ file:///C:/Users/Hp/Desktop/ocptechtrack/ocptechtrack_website/python_ldap-3.4.0-cp311-cp311-win_amd64.whl
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.6
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
